Add generic mkdir and chdir functions
The linux build broke due to the windows version of a chdir being updated and the linux version not being updated, this fixes this issue by creating the Common::Chdir and Common::Mkdir functions.

This also deletes vcruntime150_1.dll as I don't think this should be included in the repo and shouldn't even be needed due to the appvector config using a newer version of msvc now.